+++ src/Events/tests/Config/EventsConfigTest.php
@@ -5,6 +5,9 @@ declare(strict_types=1);
 namespace Spiral\Tests\Events\Config;
 
 use PHPUnit\Framework\TestCase;
+use Spiral\Core\Container\Autowire;
+use Spiral\Core\CoreInterceptorInterface;
+use Spiral\Core\CoreInterface;
 use Spiral\Events\Config\EventListener;
 use Spiral\Events\Config\EventsConfig;
 
@@ -47,5 +50,31 @@ final class EventsConfigTest extends TestCase
         $this->assertSame($listener, $config->getListeners()['foo'][1]);
         $this->assertInstanceOf(EventListener::class, $config->getListeners()['foo'][0]);
         $this->assertSame('bar', $config->getListeners()['foo'][0]->listener);
+    }
+
+    public function testGetsEmptyInterceptors(): void
+    {
+        $config = new EventsConfig();
+
+        $this->assertSame([], $config->getInterceptors());
+    }
+
+    public function testGetsInterceptors(): void
+    {
+        $config = new EventsConfig([
+            'interceptors' => [
+                'bar',
+                new class implements CoreInterceptorInterface {
+                    public function process(string $controller, string $action, array $parameters, CoreInterface $core): mixed
+                    {
+                    }
+                },
+                new Autowire('foo')
+            ]
+        ]);
+
+        $this->assertSame('bar', $config->getInterceptors()[0]);
+        $this->assertInstanceOf(CoreInterceptorInterface::class, $config->getInterceptors()[1]);
+        $this->assertInstanceOf(Autowire::class, $config->getInterceptors()[2]);
     }
 }